team: move functions out into files
